Iris work * Move some StateNotifiers to GLStateManager * Remove some redundant mixins based on sodium/angelica work and rendering state captures * Add Iris DebugScreenHandler

+++ src/main/java/com/gtnewhorizons/angelica/glsm/states/FogState.java
@@ -1,6 +1,6 @@
 package com.gtnewhorizons.angelica.glsm.states;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 import org.lwjgl.opengl.GL11;
 
 import java.nio.FloatBuffer;
@@ -8,7 +8,7 @@ import java.nio.FloatBuffer;
 public class FogState {
     public BooleanState mode = new BooleanState(GL11.GL_FOG);
     public int fogMode = GL11.GL_EXP;
-    public final Vector3f fogColor = new Vector3f(0.0F, 0.0F, 0.0F);
+    public final Vector3d fogColor = new Vector3d(0.0F, 0.0F, 0.0F);
     public float fogAlpha = 1.0F;
     public final FloatBuffer fogColorBuffer = FloatBuffer.allocate(4);
     public float density = 1.0F;
